requirements.txt
dj-database-url==0.4.1
Django==1.9.8
djangorestframework==3.4.1
whitenoise==3.2
psycopg2==2.5.4